moved middlewares
moved and renamed GitconfigAuthorizationMiddleware, NamespaceAuthorizationMiddleware and NamespaceMiddleware
